Extracting RepoMetaData
It's not really Bitbucket specific and can be reused for GitLab and
others.
