+++ source/dsl/BitBucketServerDSL.ts
@@ -1,10 +1,4 @@
-/** Key details about a repo */
-export interface RepoMetaData {
-  /** A path like "artsy/eigen" */
-  repoSlug: string
-  /** The ID for the pull/merge request "11" */
-  pullRequestID: string
-}
+import { RepoMetaData } from "./RepoMetaData"
 
 // This is `danger.bitbucket_server` inside the JSON
