+++ source/platforms/_tests/_bitbucket_server.test.ts
@@ -45,7 +45,7 @@ import { BitBucketServerAPI } from "../bitbucket_server/BitBucketServerAPI"
 import { GitCommit } from "../../dsl/Commit"
 import { FakeCI } from "../../ci_source/providers/Fake"
 import * as os from "os"
-import { RepoMetaData } from "../../dsl/BitBucketServerDSL"
+import { RepoMetaData } from "../../dsl/RepoMetaData"
 
 const EOL = os.EOL
