+++ source/platforms/_tests/_github.test.ts
@@ -57,7 +57,7 @@ import { GitHubAPI } from "../github/GitHubAPI"
 import { GitCommit } from "../../dsl/Commit"
 import { FakeCI } from "../../ci_source/providers/Fake"
 import * as os from "os"
-import { RepoMetaData } from "../../dsl/BitBucketServerDSL"
+import { RepoMetaData } from "../../dsl/RepoMetaData"
 
 const EOL = os.EOL
